Вынос scm-шага в отдельный класс
